+++ src/Conditions.php
@@ -5,7 +5,6 @@ namespace Latitude\QueryBuilder;
 
 class Conditions implements Statement
 {
-    use Traits\CanCreatePlaceholders;
     use Traits\CanUseDefaultIdentifier;
 
     /**
@@ -88,44 +87,14 @@ class Conditions implements Statement
     public function sql(Identifier $identifier = null): string
     {
         $identifier = $this->getDefaultIdentifier($identifier);
-
-        $sql = \array_reduce(
-            $this->parts,
-            function (string $sql, array $part): string {
-                if ($this->isConditions($part['condition'])) {
-                    // (...)
-                    $statement = '(' . $part['condition']->sql() . ')';
-                } else {
-                    // foo = ?
-                    $statement = $part['condition'];
-                }
-
-                if ($sql) {
-                    $statement = $part['type'] . ' ' . $statement;
-                }
-
-                return \trim($sql . ' ' . $statement);
-            },
-            ''
-        );
-
-        return $identifier->escapeExpression($sql);
+        $expression = \array_reduce($this->parts, $this->sqlReducer(), '');
+        return $identifier->escapeExpression($expression);
     }
 
     // Statement
     public function params(): array
     {
-        $reduce = function (array $params, array $part): array {
-            if ($this->isConditions($part['condition'])) {
-                return \array_merge(
-                    $params,
-                    $part['condition']->params()
-                );
-            }
-            return \array_merge($params, $part['params']);
-        };
-
-        return \array_reduce($this->parts, $reduce, []);
+        return \array_reduce($this->parts, $this->paramReducer(), []);
     }
 
     /**
@@ -148,10 +117,6 @@ class Conditions implements Statement
      */
     protected function addCondition(string $type, string $condition, array $params): self
     {
-        if ($params) {
-            $this->expandPlaceholders($condition, $params);
-        }
-
         $this->parts[] = compact('type', 'condition', 'params');
         return $this;
     }
@@ -167,26 +132,116 @@ class Conditions implements Statement
     }
 
     /**
-     * Replace placeholder with expanded placeholder for IN values.
+     * Get a function to reduce condition parts to a SQL string.
      */
-    protected function expandPlaceholders(string &$condition, array &$params)
+    protected function sqlReducer(): callable
     {
-        if ($params[0] instanceof InValue) {
-            $placeholders = $this->createPlaceholders(\count($params[0]));
-            $condition = \str_replace('?', "($placeholders)", $condition);
-            $params = $params[0]->values();
-        }
+        return function (string $sql, array $part): string {
+            if ($this->isCondition($part['condition'])) {
+                // (...)
+                $statement = "({$part['condition']->sql()})";
+            } else {
+                // foo = ?
+                $statement = $this->replaceStatementParams($part['condition'], $part['params']);
+            }
+            if ($sql) {
+                // AND ...
+                $statement = "{$part['type']} $statement";
+            }
+            return \trim($sql . ' ' . $statement);
+        };
+    }
+
+
+    /**
+     * Get a function to reduce parameters to a single list.
+     */
+    protected function paramReducer(): callable
+    {
+        return function (array $params, array $part): array {
+            if ($this->isCondition($part['condition'])) {
+                // Conditions have a parameter list already
+                return \array_merge($params, $part['condition']->params());
+            }
+            // Otherwise convert the list to a list of lists for flattening
+            return \array_merge($params, ...\array_map($this->paramLister(), $part['params']));
+        };
+    }
+
+    /**
+     * Convert all parameters to an array for flattening.
+     */
+    protected function paramLister(): callable
+    {
+        return function ($param): array {
+            if ($this->isStatement($param)) {
+                // Statements have a parameter list already
+                return $param->params();
+            }
+            // Otherwise convert to a list
+            return [$param];
+        };
     }
 
     /**
      * Check if a condition is a sub-condition.
      */
-    protected function isConditions($condition): bool
+    protected function isCondition($condition): bool
     {
         if (\is_object($condition) === false) {
             return false;
         }
-
         return $condition instanceof Conditions;
+    }
+
+    /**
+     * Check if a parameter is a statement.
+     */
+    protected function isStatement($param): bool
+    {
+        if (\is_object($param) === false) {
+            return false;
+        }
+        return $param instanceof Statement;
+    }
+
+    /**
+     * Check if any parameter is a statement.
+     */
+    protected function hasStatementParam(array $params): bool
+    {
+        foreach ($params as $param) {
+            if ($this->isStatement($param)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Replacement statement parameters with SQL expression.
+     */
+    protected function replaceStatementParams(string $statement, array $params): string
+    {
+        if ($this->hasStatementParam($params) === false) {
+            return $statement;
+        }
+        // Maintain an offset position, as preg_replace_callback() does not provide one
+        $index = 0;
+        return \preg_replace_callback('/\?/', function ($matches) use (&$index, $params) {
+            try {
+                if ($this->isStatement($params[$index])) {
+                    // Replace any statement placeholder with the generated SQL
+                    return $params[$index]->sql();
+                } else {
+                    // And leave all other placeholders intact
+                    return $matches[0];
+                }
+            } finally {
+                // This funky usage of finally allows us to increment the offset
+                // after all other code in the block has been executed.
+                $index++;
+            }
+        }, $statement);
     }
 }