+++ src/InsertQuery.php
@@ -3,18 +3,18 @@ declare(strict_types=1);
 
 namespace Latitude\QueryBuilder;
 
+use InvalidArgumentException;
 use Iterator;
 
 class InsertQuery implements Statement
 {
     use Traits\CanConvertIteratorToString;
-    use Traits\CanReplaceBooleanAndNullValues;
     use Traits\CanUseDefaultIdentifier;
 
     /**
      * Create a new insert query.
      */
-    public static function make(string $table, array $map): InsertQuery
+    public static function make(string $table, array $map = []): InsertQuery
     {
         $query = new static();
         $query->table($table);
@@ -34,12 +34,40 @@ class InsertQuery implements Statement
     }
 
     /**
+     * Set the columns to insert.
+     */
+    public function columns(string ...$columns): self
+    {
+        $this->columns = $columns;
+        return $this;
+    }
+
+    /**
+     * Append values to insert.
+     */
+    public function values(...$values): self
+    {
+        if (\count($values) !== \count($this->columns)) {
+            throw new InvalidArgumentException(sprintf(
+                'Number of values (%d) does not match number of columns (%d)',
+                \count($values),
+                \count($this->columns)
+            ));
+        }
+        $this->values[] = ValueList::make(...$values);
+        return $this;
+    }
+
+    /**
      * Set the columns and values to insert.
+     *
+     * NOTE: Existing values will be replaced!
      */
     public function map(array $map): self
     {
-        $this->columns = \array_keys($map);
-        $this->params = \array_values($map);
+        $this->values = [];
+        $this->columns(...\array_keys($map));
+        $this->values(...\array_values($map));
         return $this;
     }
 
@@ -49,17 +77,18 @@ class InsertQuery implements Statement
         $identifier = $this->getDefaultIdentifier($identifier);
 
         return \sprintf(
-            'INSERT INTO %s (%s) VALUES (%s)',
+            'INSERT INTO %s (%s) VALUES %s',
             $identifier->escape($this->table),
             \implode(', ', $identifier->all($this->columns)),
-            $this->stringifyIterator($this->generatePlaceholders())
+            $this->stringifyIterator($this->insertLines())
         );
     }
 
     // Statement
     public function params(): array
     {
-        return $this->placeholderParams();
+        // [[a], [b], [c]] -> [a, b, c]
+        return \array_merge(...\array_map($this->paramLister(), $this->values));
     }
 
     /**
@@ -73,17 +102,27 @@ class InsertQuery implements Statement
     protected $columns = [];
 
     /**
-     * @var array
+     * @var ValueList[]
      */
-    protected $params = [];
+    protected $values = [];
 
     /**
-     * Generate a placeholder.
+     * Generate a list of insert lines.
      */
-    protected function generatePlaceholders(): Iterator
+    protected function insertLines(): Iterator
     {
-        foreach (\array_keys($this->params) as $index) {
-            yield $this->placeholderValue($index);
+        foreach ($this->values as $line) {
+            yield "({$line->sql()})";
         }
+    }
+
+    /**
+     * Convert all parameters to an array for flattening.
+     */
+    protected function paramLister(): callable
+    {
+        return function (ValueList $values): array {
+            return $values->params();
+        };
     }
 }