+++ src/ValueList.php
@@ -3,9 +3,12 @@ declare(strict_types=1);
 
 namespace Latitude\QueryBuilder;
 
+use Countable;
 use Iterator;
 
-class ValueList implements Statement
+class ValueList implements
+    Countable,
+    Statement
 {
     use Traits\CanConvertIteratorToString;
     use Traits\CanReplaceBooleanAndNullValues;
@@ -20,10 +23,16 @@ class ValueList implements Statement
         return $values;
     }
 
+    // Countable
+    public function count()
+    {
+        return \count($this->params);
+    }
+
     // Statement
     public function sql(Identifier $identifier = null): string
     {
-        return '(' . $this->stringifyIterator($this->generatePlaceholders()) . ')';
+        return $this->stringifyIterator($this->generatePlaceholders());
     }
 
     // Statement