+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0]
+
+### Changed
+
+- Statements are now supported as parameters in conditions
+- Multi-line inserts are now supported by `InsertQuery`
+
+### Removed
+
+- `InValue` has been removed in favor of `ValueList`
+- `InsertMultipleQuery` has been removed
+
 ## [1.4.0]
 
 ### Added